chore: combine all multisig tools in one
