+++ src/tools/index.ts
@@ -3,13 +3,6 @@ export * from "./close_empty_token_accounts";
 export * from "./create_3land_collectible";
 export * from "./create_gibwork_task";
 export * from "./create_image";
-export * from "./create_multisig";
-export * from "./multisig_approve_proposal";
-export * from "./multisig_create_proposal";
-export * from "./multisig_deposit_to_treasury";
-export * from "./multisig_execute_proposal";
-export * from "./multisig_reject_proposal";
-export * from "./multisig_transfer_from_treasury";
 export * from "./create_tiplinks";
 export * from "./deploy_collection";
 export * from "./deploy_token";
@@ -31,6 +24,7 @@ export * from "./launch_pumpfun_token";
 export * from "./lend";
 export * from "./manifest_trade";
 export * from "./mint_nft";
+export * from "./multisig_tools";
 export * from "./openbook_create_market";
 export * from "./orca_close_position";
 export * from "./orca_create_clmm";
